Use int64_t for num_tokens to prevent offset overflow
Change num_tokens from int to int64_t in the kernel signature, launcher,
and all offset calculations (baseOffset, v_output_offset, inputOffset,
outputBase/outputOffset, quantize_store_fp8 offset parameter).

This prevents integer overflow when batch_size * seq_len * num_heads *
head_dim exceeds 2^31 (e.g. batch=64 with seq_len=7680 and 24 heads).

AI-assisted.

Made-with: Cursor

+++ csrc/norm.cu
@@ -304,7 +304,7 @@ void fused_qk_rmsnorm_rope_run(TensorView qkv_in, TensorView q_weight, TensorVie
 
   launchFusedQKNormRope(
       qkv_in.data_ptr(), q_out.data_ptr(), k_out.data_ptr(), v_out.data_ptr(),
-      static_cast<int>(num_tokens), static_cast<int>(seq_len), static_cast<int>(ppf),
+      num_tokens, static_cast<int>(seq_len), static_cast<int>(ppf),
       static_cast<int>(pph), static_cast<int>(ppw), static_cast<int>(num_frame_channels),
       static_cast<int>(num_height_channels), static_cast<int>(num_width_channels),
       static_cast<int>(num_heads_q), static_cast<int>(num_heads_k), static_cast<int>(num_heads_v),
